[ALPS02968842] cmdq: Enable 64bit handle
[Detail]
Change handle and PA to 64 bit.

MTK-Commit-Id: 10cf01cdd2b4e53b71e1b9301f2e166d3dcb3bbc

CR-Id: ALPS02968842
Feature: CMDQ

+++ drivers/misc/mediatek/cmdq/v2/cmdq_record.h
@@ -182,12 +182,12 @@ extern "C" {
  *     support only when secure OS enabled
  */
 	int32_t cmdq_op_write_reg_secure(cmdqRecHandle handle, uint32_t addr,
-				   CMDQ_SEC_ADDR_METADATA_TYPE type, uint32_t baseHandle,
+				   CMDQ_SEC_ADDR_METADATA_TYPE type, uint64_t baseHandle,
 				   uint32_t offset, uint32_t size, uint32_t port);
 	int32_t cmdqRecWriteSecure(cmdqRecHandle handle,
 				   uint32_t addr,
 				   CMDQ_SEC_ADDR_METADATA_TYPE type,
-				   uint32_t baseHandle,
+				   uint64_t baseHandle,
 				   uint32_t offset, uint32_t size, uint32_t port);
 
 /* tablet use */
